requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
imutils==0.5.4
isodate==0.6.1
lxml==4.8.0
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
pycountry==22.3.5
pycryptodome==3.14.1
pyparsing==3.0.8
PySocks==1.7.1
pytesseract==0.3.9
requests==2.27.1
six==1.16.0
streamlink==4.0.1
tesserocr==2.5.2
urllib3==1.26.9
websocket-client==1.3.2
